tests/wpt/web-platform-tests/app-history/navigate/state/updateCurrent/state-cross-document-restore.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    i.contentWindow.appHistory.updateCurrent({ state: { data: "value" } });
    assert_equals(i.contentWindow.appHistory.entries().length, 1);
    assert_equals(i.contentWindow.appHistory.current.getState().data, "value");

    let navigated_back = false;
    i.contentWindow.appHistory.navigate("?1");
    i.onload = t.step_func(() => {
      if (navigated_back) {
        let back_entry = i.contentWindow.appHistory.current;
        assert_equals(i.contentWindow.appHistory.entries().length, 2);
        assert_equals(back_entry.index, 0);
        assert_equals(back_entry.getState().data, "value");
        t.done();
      } else {
        assert_equals(i.contentWindow.appHistory.entries().length, 2);
        assert_equals(i.contentWindow.appHistory.current, i.contentWindow.appHistory.entries()[1]);
        assert_equals(i.contentWindow.appHistory.current.getState(), undefined);
        history.back();
        navigated_back = true;
      }
    });
  });
}, "appHistoryEntry.getState() behavior after navigating away and back");
</script>
